scripts/requirements.txt
# Claude API への直接呼び出しに httpx を使用（anthropic SDK は不使用）
httpx>=0.27.0,<1.0.0
Pillow>=10.0.0,<12.0.0
tweepy>=4.14.0,<5.0.0
openpyxl>=3.1.0,<4.0.0
pdfplumber>=0.11.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
